[tests] Move base test cases to tests.base
